+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.3.0
+ - add node['jlsolrcloud]['remote_jmx'] attribute to enable remote JMX
+ - more complete chefspec coverage for jlsolrcloud::configure
+
 ### 0.2.2
  - append console output to file using >> instead of > to make logrotate
    copytruncate work